requirements.txt
numpy
pixiv-api
Pillow
beautifulsoup4
tkinterdnd2
